seperate bundling from server code
