+++ src/script/webpack/assetUrl.spec.js
